Specify the set of well known HTTP Content-Type values as AWS expects them.
AWS expects these content type headers to set exactly right,
as they are used in the calculation of checksums and cryptograph signatures
That means they need to precise on a binary level, including precise
case sensetivity and formatting / white space etc.
Since HTTP does not define it as such in any standard, AKKA Http "mangels"
the values as far as AWS is concerned.
To mitigate this behavior, this map contains precise representations of
the most known content types that AWS expects and how it expects them.

This fixes issues where request signature calculated by the server do not match
the signature provided by the client.